Rename provider 'ovirt' to 'olvm'
